Create rocket or missile horizontal range calculator.py
hi, everyone i am shaikybombs and a 15year old programmer who has started to learn python and I am glad that  my first project is regarding the space and rocket technology here the code is written in the form such that projectile motion formuales are expressed in such a way that we can calculate the needed values such as initial velocity, acceleration and force required and time of flight for missiles to land on designated target. this project is helpful in understanding the projectile motion and can give you accurate values for the required quantities by which you can make a rocket. if you want a vertical range calculator please let me know.
shaikybombs signing out.